Add suppport for MDA ports compatibility
The VGA CRT register (and other registers) is accessed by a port that
depends on the status of the miscellaneous output register. Before the
change, we used to set it to CGA compatibility mode. Now a subroutine to
change mode is implemented, and the cursor-moving subroutine uses the
current mode instead of CGA mode.