Split kernel table and io table
It is probably better to split the two tables, since the subroutines and
information they provide are located in two different executable, and
they are conceptually different. This is also useful in a possible (iven
if unlikely) port to a new architecture, where the kernel subroutines
and informaation would remain the same, while the low level io could
change. I have also put the memory detection code in the kernel file,
since it isn't strictly io, therefore it probably makes more sense
there.