Start implementing keyboard interrupts
I decided that the keyboard driver will operate in two different modes
(buffered and unbuffered) as explained in the io/keyb.s file. The first
one to be implemented will be buffered mode: unbuffered is not a
priority, since is not very useful for console applications.
I also implemented a first subroutine to let the user get raw scancodes in
buffered mode.